- scaladocs updates - config value renaming
