+++ colossus-metrics/src/main/resources/reference.conf
@@ -1,8 +1,8 @@
 colossus {
   metrics {
     collect-system-metrics: true
-    metric-intervals: ["1 second", "1 minute"]
-    metric-address: "/"
+    collection-intervals: ["1 second", "1 minute"]
+    namespace: "/"
     collectors-defaults {
       rate {
         prune-empty: false
@@ -16,5 +16,9 @@ colossus {
 
       }
     }
+    ##Metrics can be listed here as well:
+    # /my/path/to/my/histogram {
+     # ...configuration
+    ##}
   }
 }
