+++ colossus-metrics/src/main/scala/colossus/metrics/MetricSystem.scala
@@ -28,29 +28,43 @@ class MetricInterval private[metrics](val namespace : MetricAddress,
 }
 
 /**
- * The MetricSystem is a set of actors which handle the background operations of dealing with metrics. In most cases,
- * you only want to have one MetricSystem per application.
- *
- * Metrics are generated periodically by a Tick message published on the global event bus. By default this happens once
- * per second, but it can be configured to any time interval. So while events are being collected as they occur,
- * compiled metrics (such as rates and histogram percentiles) are generated once per tick.
- *
- * @param namespace the base of the url describing the location of metrics within the system
- * @param metricIntervals Map of all MetricIntervals for which this system collects Metrics.
- * @param config Config object from which Metric configurations will be sourced.
+  * The MetricSystem is a set of actors which handle the background operations of dealing with metrics. In most cases,
+  * you only want to have one MetricSystem per application.
+  *
+  * The currently provided metric types are [[colossus.metrics.Rate]], [[colossus.metrics.Histogram]] and [[colossus.metrics.Counter]].
+  * New metric types can be created by implementing the [[colossus.metrics.Collector]] trait.
+  *
+  * Namespace is the root of this MetricSystem's addressing.  All metrics which are registered within this MetricSystem will have their
+  * addresses prefixed with this value.
+  *
+  * Metrics are collected and reported for each collectionInterval specified.
+  *
+  * Metric Configuration
+  *
+  * A MetricSystem's configuration contains defaults for each metric type.  It can also contain configuration for individual metrics.
+  * When a new Metric is created, a MetricSystem will first look for a metric entry corresponding to the MetricAddress of the metric.
+  * If it finds one, it will use that entry overlaid on the default entry for that metric type to generate a complete configuration.
+  * If it does not find one, it will fallback to just using the default entry for that metric type.
+  *
+  * @param namespace Base url for all metrics within this MetricSystem
+  * @param collectionIntervals Intervals for which this MetricSystem reports its data.
+  * @param config Config object from which Metric configurations will be sourced.
  */
-case class MetricSystem private[metrics] (namespace: MetricAddress, metricIntervals : Map[FiniteDuration, MetricInterval],
+case class MetricSystem private[metrics] (namespace: MetricAddress, collectionIntervals : Map[FiniteDuration, MetricInterval],
                                           collectionSystemMetrics : Boolean, config : Config) {
 
-  implicit val base = new Collection(CollectorConfig(metricIntervals.keys.toSeq, config))
+  implicit val base = new Collection(CollectorConfig(collectionIntervals.keys.toSeq, config))
   registerCollection(base)
 
   def registerCollection(collection: Collection): Unit = {
-    metricIntervals.values.foreach(_.intervalAggregator ! IntervalAggregator.RegisterCollection(collection))
+    collectionIntervals.values.foreach(_.intervalAggregator ! IntervalAggregator.RegisterCollection(collection))
   }
 
 }
 
+/**
+  * Factory for [[colossus.metrics.MetricSystem]] instances
+  */
 object MetricSystem {
 
   val ConfigRoot = "colossus.metrics"
@@ -58,22 +72,22 @@ object MetricSystem {
   /**
    * Constructs a metric system
    *
-   * @param namespace the base of the url describing the location of metrics within the system
-   * @param metricIntervals How often to report metrics
+   * @param namespace Base url for all metrics within this MetricSystem
+   * @param collectionIntervals How often to report metrics
    * @param collectSystemMetrics whether to collect metrics from the system as well
-   * @param config Typesafe Config source which shoudl contain configuration for a MetricSystem.  This Configuration object is used during
+   * @param config Typesafe Config source which should contain configuration for a MetricSystem.  This Configuration object is used during
    *               Collector(Rate, Histogram specifically) creation.  If this Config does not have the corresponding colossus.metrics.collectors.defaults for those
    *               then Collector creation will fail if it tries to reference them.
    * @param system the actor system the metric system should use
    * @return
    */
-  def apply(namespace: MetricAddress, metricIntervals: Seq[FiniteDuration] = Seq(1.second, 1.minute),
+  def apply(namespace: MetricAddress, collectionIntervals: Seq[FiniteDuration] = Seq(1.second, 1.minute),
             collectSystemMetrics: Boolean = true, config : Config = ConfigFactory.defaultReference())
   (implicit system: ActorSystem): MetricSystem = {
     import system.dispatcher
 
 
-    val m : Map[FiniteDuration, MetricInterval] = metricIntervals.map{ interval =>
+    val m : Map[FiniteDuration, MetricInterval] = collectionIntervals.map{ interval =>
       val snap = Agent[MetricMap](Map())
       val aggregator : ActorRef = createIntervalAggregator(system, namespace, interval, snap, collectSystemMetrics)
       val i = new MetricInterval(namespace, interval, aggregator, snap)
@@ -88,13 +102,18 @@ object MetricSystem {
     system.actorOf(Props(classOf[IntervalAggregator], namespace, interval, snap, collectSystemMetrics))
   }
 
+  /**
+    * Create a system which does nothing.  Useful for testing/debugging.
+    * @param system
+    * @return
+    */
   def deadSystem(implicit system: ActorSystem) = {
     MetricSystem(Root / "DEAD", Map[FiniteDuration, MetricInterval](), false, ConfigFactory.defaultReference())
   }
 
   /**
     * Create a new MetricSystem, using only the defaults provided by the corresponding "colossus.metrics" config path.
-    * A Config object will be created via "ConfigFactory.load()"
+    * A Config object will be created via {{{ConfigFactory.load()}}}
     * @param system
     * @return
     */
@@ -105,7 +124,7 @@ object MetricSystem {
   /**
     * Create a new MetricSystem using the supplied configPath.  This configPath will be overlaid on top of the default "colossus.metrics"
     * config path.
-    * A Config object will be created via "ConfigFactory.load()"
+    * A Config object will be created via {{{ConfigFactory.load()}}}
     *
     * @param configPath  The path to the configuration.
     * @param system
@@ -136,8 +155,8 @@ object MetricSystem {
     val mergedConfig = config.withValue(ConfigRoot, metricsObject)
 
     val collectSystemMetrics = metricsConfig.getBoolean("collect-system-metrics")
-    val metricIntervals = metricsConfig.getFiniteDurations("metric-intervals")
-    val metricAddress = metricsConfig.getString("metric-address")
+    val metricIntervals = metricsConfig.getFiniteDurations("collection-intervals")
+    val metricAddress = metricsConfig.getString("namespace")
     MetricSystem(MetricAddress(metricAddress), metricIntervals, collectSystemMetrics, mergedConfig)
   }
 }
