+++ colossus/src/main/resources/reference.conf
@@ -101,15 +101,3 @@ metric-system{
 io-system{
 
 }
-
-
-
-imageReg{
-
-  metrics {
-    my-rate{
-      type : "Rate"
-    }
-  }
-
-}
